# docs/swagger.yaml
basePath: /api
definitions:
...
  agent.HostExecRequest:
    properties:
      command:
        description: shell -c argument; the whole string is one command line
        type: string
      timeout:
        description: seconds; default 30, max 300
        type: integer
    type: object
  agent.HostExecResponse:
    properties:
      duration_ms:
        type: integer
      exit_code:
        type: integer
      node:
        type: string
      stderr:
        type: string
      stdout:
        type: string
      timed_out:
        type: boolean
    type: object
...
info:
  contact:
    name: Jetty Support
    url: https://github.com/ncwardell/jetty
  description: P2P Docker Compose orchestration with Cloudflare WARP mesh networking
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Jetty API
  version: 0.0.x
paths:
...
  /host/exec:
    post:
      consumes:
      - application/json
      description: One-shot exec. Requires JETTY_HOST_SHELL=true on the target agent
        and AdminKey auth. ?node=<id|name> proxies to a peer.
      parameters:
      - description: Peer ID or name; defaults to this node
        in: query
        name: node
        type: string
      - description: Command to run
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/agent.HostExecRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agent.HostExecResponse'
      summary: Run one command on the host and return its output
      tags:
      - terminal
  /host/shell:
    get:
      description: 'Upgrades to a WebSocket and runs an interactive shell. With --pid=host
        on the docker run, this is a true host shell via nsenter. Without it, the
        shell runs inside the agent container and a banner explains the limitation.
        DISABLED by default. Set JETTY_HOST_SHELL=true to enable. Auth: ?api_key=<secret>
        or X-API-Key header.'
      parameters:
      - description: Shell to run (default /bin/bash, falls back to /bin/sh)
        in: query
        name: shell
        type: string
      responses: {}
      summary: Open a shell on the host (WebSocket)
      tags:
      - terminal
...
  /workloads/{name}/exec:
    get:
      description: 'Upgrades to a WebSocket and runs docker exec -i -t against the
        workload''s first container (or ?service=X for a specific service). Use ?shell=/bin/bash
        to override the default /bin/sh. Auth: ?api_key=<secret> or X-API-Key header.'
      parameters:
      - description: Workload name
        in: path
        name: name
        required: true
        type: string
      - description: 'Compose service name (default: first container)'
        in: query
        name: service
        type: string
      - description: Shell to exec (default /bin/sh)
        in: query
        name: shell
        type: string
      responses: {}
      summary: Open an interactive shell in a workload container (WebSocket)
      tags:
      - terminal
...
schemes:
- https
- http
security:
- ApiKeyAuth: []
securityDefinitions:
  ApiKeyAuth:
    description: API key for authentication (JETTY_SECRET)
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"